Extensible runvmtest tool
Add the ability to configure extracting more stuff from containers and
assign it to env vars.

For example, u-root VM tests also take a multiboot kernel and an EDK2
build as a dependency. This allows providing these dependencies via
containers and extra env vars.

Removes the ability to configure overrides via the command-line.

Signed-off-by: Chris Koch <chrisko@google.com>